octopi/requirements.txt
numpy
opencv-python
pillow
torch
tqdm
natsort
torchvision
transformers==4.45.1
peft==0.5.0
python-dotenv
sentencepiece
protobuf
bitsandbytes
wandb
pandas
matplotlib
seaborn
fastapi
openai
scikit-learn
qwen_vl_utils
jupyter
ipykernel